pages/requirements.txt
pandas
statistics
statsmodel
scipy
numpy